requirements.txt
django==5.1.2
djangorestframework==3.15.2
flake8==7.1.1